+++ level2/AreaOfTriangle.java
@@ -6,9 +6,9 @@ public class AreaOfTriangle {
 		double height;
 		Scanner input = new Scanner(System.in);
 		System.out.println("Enter base of triangle(in cm): ");
-		base = input.nextInt();
+		base = input.nextDouble();
 		System.out.println("Enter height of triangle(in cm): ");
-		height = input.nextInt();
+		height = input.nextDouble();
 		
 		double inchFactor = 2.54;//1 inch = 2.54cm
 		double heightInInches = height/2.54;
