+++ level2/quotientAndRemainder.java
@@ -6,8 +6,8 @@ public class quotientAndRemainder {
 		int number2;
 		Scanner input = new Scanner(System.in);
 		System.out.println("Enter two numbers : ");
-		number1 = input.nextInt();
-		number2 = input.nextInt();
+		number1 = input.nextDouble();
+		number2 = input.nextDouble();
 		int quotient = number1/number2;
 		int remainder = number1%number2;
