+++ level2/sideOfSqr.java
@@ -6,7 +6,7 @@ public class sideOfSqr {
 		double side;
 		Scanner input = new Scanner(System.in);
 		System.out.println("Enter perimeter of square : ");
-		perimeter = input.nextInt();
+		perimeter = input.nextDouble();
 		side = perimeter/4;
 		System.out.println("Side of square = " + side);
 	}
