+++ level2/totalPrice.java
@@ -6,9 +6,9 @@ public class totalPrice {
 		double quantity;
 		Scanner input = new Scanner(System.in);
 		System.out.println("Enter price of item: ");
-		price = input.nextInt();
+		price = input.nextDouble();
 		System.out.println("Enter quantity of item: ");
-		quantity = input.nextInt();
+		quantity = input.nextDouble();
 		
 		System.out.println("The total purchase price is INR  : " + (price*quantity) + " if the quantity " + quantity);
 	}
